requirements.txt
flask
flask-pymongo
flask_socketio

google-cloud-language
eventlet
